requirements/requirements-cpu.txt
imageai==3.0.3
tensorflow-cpu==2.9.0
opencv-python-headless==4.6.0.66
pillow>=9.0.0
numpy>=1.21.0
pyyaml>=6.0